Fix language generation for Forums and (indirectly) Posts. Fixes #349
The old versions (including the legacy Hadoop Datagen) had a bug where
the language of a Forum was selected as a random number between 0 and
the number of languages a person spoke (0, 1, or 2).

The fix is trivial: this random number should be used to *select* a
language from the list of languages spoken by the person (instead of
simply using this number as the id of the language).